refactor: prepares template for using other services
