+++ src/Application/Gateway.Application/Extensions/ServiceCollectionExtensions.cs
@@ -1,4 +1,6 @@
-using Gateway.Application.Contracts;
+using Gateway.Application.Contracts.GrpcClients;
+using Gateway.Application.GrpcClients;
+using Gateway.Application.Options;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 
@@ -10,12 +12,15 @@ public static class ServiceCollectionExtensions
     {
         collection.AddGrpcClient<Users.UsersService.Contracts.UsersService.UsersServiceClient>((services, options) =>
         {
-        GrpcClientOptions grpcServerOption = services.GetRequiredService<IOptions<GrpcClientOptions>>().Value;
-        options.Address = new Uri(grpcServerOption.BaseAddress);
+            GrpcServerOptions grpcServerOptions = services.GetRequiredService<IOptions<GrpcServerOptions>>().Value;
+            if (grpcServerOptions.UserService is null)
+                throw new NullReferenceException("Options object for GrpcUserClient is null");
+            options.Address = new Uri(grpcServerOptions.UserService.BaseAddress);
         });
 
-        collection.AddScoped<IGrpcClient, GrpcClient>();
+        collection.AddScoped<IGrpcUserClient, GrpcUserClient>();
 
+        // TODO. Add other clients and their options
         return collection;
     }
 }
