--- src/Application/Gateway.Application/GrpcClientOptions.cs
@@ -1,6 +1,0 @@
-namespace Gateway.Application;
-
-public class GrpcClientOptions
-{
-    public string BaseAddress { get; set; } = string.Empty;
-}
