+++ src/Application/Gateway.Application/Options/GrpcServerOptions.cs
@@ -1,0 +1,10 @@
+namespace Gateway.Application.Options;
+
+public class GrpcServerOptions
+{
+    public ServiceOptions? UserService { get; set; }
+
+    public ServiceOptions? AccommodationService { get; set; }
+
+    public ServiceOptions? BookingService { get; set; }
+}
