+++ src/Application/Gateway.Application/Options/ServiceOptions.cs
@@ -1,0 +1,6 @@
+namespace Gateway.Application.Options;
+
+public class ServiceOptions
+{
+    public string BaseAddress { get; set; } = string.Empty;
+}
