+++ src/Gateway/appsettings.json
@@ -1,27 +1,5 @@
 {
   "Application": {},
-  "Infrastructure": {
-    "Persistence": {
-      "Postgres": {
-        "Host": "",
-        "Port": -1,
-        "Database": "",
-        "Username": "",
-        "Password": "",
-        "SslMode": "Prefer",
-        "Pooling": true
-      }
-    }
-  },
-  "Presentation": {
-    "Kafka": {
-      "Host": "",
-      "Consumers": {
-      },
-      "Producers": {
-      }
-    }
-  },
   "AllowedHosts": "*",
   "Kestrel": {
     "Http": {
@@ -29,8 +7,16 @@
       "Protocols": "Http1"
     }
   },
-  "GrpcClient": {
-    "BaseAddress": "http://localhost:8070"
+  "GrpcServer": {
+    "UserService" : {
+      "BaseAddress": "http://localhost:8070"
+    },
+    "AccommodationService" : {
+      "BaseAddress": ""
+    },
+    "BookingService" : {
+      "BaseAddress": ""
+    }
   },
   "Platform": {
     "ServiceName": "<>",
